// ui/src/generated-api-client/models/task-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * UPOU TODO API
 * Task management API for UPOU students and faculty
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { PriorityEnum } from "./priority-enum";

export interface TaskRequest {
  title: string;
  description?: string;
  dtstart?: string | null;
  rrule?: string | null;
  anchor_mode?: TaskRequestAnchorModeEnum | null;
  priority?: PriorityEnum;
  tags?: Array<string>;
  completion_date?: string | null;
  order?: number;
  section: number;
  above_task?: number;
  below_task?: number;
  source_section?: number;
}

export const TaskRequestAnchorModeEnum = {
  Scheduled: "SCHEDULED",
  Completed: "COMPLETED",
  Empty: "",
} as const;

export type TaskRequestAnchorModeEnum =
  (typeof TaskRequestAnchorModeEnum)[keyof typeof TaskRequestAnchorModeEnum];